// src/app/models/ProductFinalModel.ts
export class ProductFinalModel{
    productId: number
    userId: string;
    departmentId: number;
    productName: string;
    productPrice: number;
    productQuantity: number;
    productSize: string;
    productDescription: string;
    forswap: boolean;
    forsell: boolean;
    categoryId:number;
    
    // added
    
    ownerFirstName: string;
    ownerLastName: string;
    sInCart: string;
    sInFav: string;
    sIsOwner: string;
    email: string;
    productImage: string;
    e: string;
    
}